+++ src/frontend/js/pages/TeacherCourseDashboardLoader/CourseRunList/index.spec.tsx
@@ -2,6 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { IntlProvider } from 'react-intl';
 import { CunninghamProvider } from '@openfun/cunningham-react';
 import { capitalize } from 'lodash-es';
+import { MemoryRouter } from 'react-router-dom';
 import { CourseRunFactory } from 'utils/test/factories/joanie';
 import CourseRunList from '.';
 
@@ -11,13 +12,15 @@ describe('pages/TeacherCourseDashboardLoader/CourseRunList', () => {
     render(
       <IntlProvider locale="en">
         <CunninghamProvider>
-          <CourseRunList courseRuns={courseRuns} />
+          <MemoryRouter>
+            <CourseRunList courseRuns={courseRuns} />
+          </MemoryRouter>
         </CunninghamProvider>
       </IntlProvider>,
     );
     const [courseRunOne, courseRunTwo] = courseRuns;
     expect(screen.getByTitle(capitalize(courseRunOne.title))).toBeInTheDocument();
     expect(screen.getByTitle(capitalize(courseRunTwo.title))).toBeInTheDocument();
-    expect(screen.getAllByRole('button', { name: 'go to classroom' }).length).toEqual(2);
+    expect(screen.getAllByRole('link', { name: 'go to classroom' }).length).toEqual(2);
   });
 });